- added basic points system 	- when you enter a country, you receive a point
