+++ src/services/i18n/lang/es.json
@@ -33,11 +33,11 @@
     },
     "guess-input": {
       "type-country-name": "Escribe el nombre de un país...",
-      "country-duplicated": "{country} ya fue ingresado"
+      "country-duplicated": "{country} ya fue ingresado",
+      "restart-game": "Reiniciar juego"
     },
     "timer-section": {
-      "no-time": "Sin tiempo",
-      "restart-game": "Reiniciar juego"
+      "no-time": "Sin tiempo"
     },
     "country-flag": {
       "flag": "Bandera de {country}"
@@ -53,6 +53,7 @@
     }
   },
   "common": {
+    "points": "0 puntos | 1 punto | {count} puntos",
     "restart": "Reiniciar",
     "cancel": "Cancelar",
     "new-game": "Nuevo juego",
